# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chronivaro REST API
  description: |
    REST API specification for Chronivaro, a Resource-Order-Activity time tracking and absence management system built on Strolch.
    Provides endpoints for time tracking, timers, presence overview, absence requests, period closing, reports, and administrative management.
  version: 0.1.0-SNAPSHOT
  contact:
    name: Chronivaro
    url: https://github.com/eitch/Chronivaro
servers:
  - url: /
    description: Current server instance
paths:
...
  /chronivaro/v1/admin/employees/{id}/schedules/{scheduleId}:
    get:
      summary: Get employee schedule by ID
      description: Returns a specific schedule period for an employee.
      operationId: getEmployeeScheduleById
      tags:
        - Admin - Employees
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Employee identifier
          schema:
            type: string
        - name: scheduleId
          in: path
          required: true
          description: Schedule identifier
          schema:
            type: string
      responses:
        '200':
          description: Schedule retrieved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update employee schedule
      description: Updates schedule weekday hours and validity range. Supports optimistic concurrency via `If-Match`.
      operationId: updateEmployeeSchedule
      tags:
        - Admin - Employees
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Employee identifier
          schema:
            type: string
        - name: scheduleId
          in: path
          required: true
          description: Schedule identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpdateDto'
      responses:
        '200':
          description: Schedule updated successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Remove employee schedule
      description: Removes a schedule interval. Supports optimistic concurrency via `If-Match`.
      operationId: removeEmployeeSchedule
      tags:
        - Admin - Employees
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Employee identifier
          schema:
            type: string
        - name: scheduleId
          in: path
          required: true
          description: Schedule identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Schedule removed successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chronivaro/v1/admin/schedule-templates:
    get:
      summary: List schedule templates
      description: Returns predefined weekly schedule templates (e.g., 100% Full Time 40h, 80% 32h).
      operationId: getScheduleTemplates
      tags:
        - Admin - Schedule Templates
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Schedule templates retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ScheduleTemplateDto'
                  - $ref: '#/components/schemas/PagedScheduleTemplateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create schedule template
      description: Creates a new reusable weekly schedule model.
      operationId: createScheduleTemplate
      tags:
        - Admin - Schedule Templates
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleTemplateCreateDto'
      responses:
        '200':
          description: Schedule template created successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleTemplateDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/admin/schedule-templates/{id}:
    get:
      summary: Get schedule template by ID
      description: Returns details of a specific schedule template.
      operationId: getScheduleTemplateById
      tags:
        - Admin - Schedule Templates
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule template identifier
          schema:
            type: string
      responses:
        '200':
          description: Schedule template retrieved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleTemplateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update schedule template
      description: Updates template name and day target allocations. Supports optimistic concurrency via `If-Match`.
      operationId: updateScheduleTemplate
      tags:
        - Admin - Schedule Templates
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule template identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleTemplateCreateDto'
      responses:
        '200':
          description: Schedule template updated successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleTemplateDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Remove schedule template
      description: Removes a schedule template. Supports optimistic concurrency via `If-Match`.
      operationId: removeScheduleTemplate
      tags:
        - Admin - Schedule Templates
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule template identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Schedule template removed successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # Admin - Users
  # ==========================================
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: strolch.authorization
      description: Strolch session cookie containing the user authentication token.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token containing the Strolch user certificate ID.
  headers:
    ETag:
      description: Entity version validator for optimistic concurrency control (e.g., W/"1" or "1").
      schema:
        type: string
        example: '"1"'
...
    X-Correlation-Id:
      description: Unique identifier for tracking and correlating requests across logs and client telemetry.
      schema:
        type: string
        format: uuid
        example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Zero-based item offset for pagination (defaults to 0).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
        example: 0
    LimitParam:
      name: limit
      in: query
      description: Number of items to return (defaults to 50, maximum 1000).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
        example: 50
    IfMatchParam:
      name: If-Match
      in: header
      required: false
      description: Target entity version for optimistic concurrency control.
      schema:
        type: string
        example: '"1"'
  responses:
    BadRequest:
      description: Bad request - Validation failure or malformed payload
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Unauthorized:
      description: Unauthorized - Authentication is missing or invalid
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Forbidden:
      description: Forbidden - Insufficient user privileges or role permissions
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    NotFound:
      description: Not found - Requested resource does not exist
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Conflict:
      description: Conflict - Concurrency mismatch (stale If-Match) or business state violation
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    InternalServerError:
      description: Internal server error - Unexpected system failure
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
  schemas:
...
    ErrorDto:
      type: object
      required:
        - error
        - code
        - message
        - correlationId
      properties:
        error:
          type: boolean
          example: true
        code:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Invalid input parameters'
        correlationId:
          type: string
          example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    FieldErrorDto:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: 'email'
        message:
          type: string
          example: 'Email address is invalid'
    SuccessResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation completed successfully'
...
    ScheduleDto:
      type: object
      properties:
        id:
          type: string
          example: 'sch-001'
        employeeId:
          type: string
          example: 'emp-001'
        validFrom:
          type: string
          format: date-time
          example: '2026-01-01T00:00:00+01:00'
        validTo:
          type: string
          format: date-time
          nullable: true
          example: '2026-12-31T23:59:59+01:00'
        monday:
          type: integer
          example: 480
        tuesday:
          type: integer
          example: 480
        wednesday:
          type: integer
          example: 480
        thursday:
          type: integer
          example: 480
        friday:
          type: integer
          example: 480
        saturday:
          type: integer
          example: 0
        sunday:
          type: integer
          example: 0
...
    ScheduleUpdateDto:
      type: object
      required:
        - validFrom
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
      properties:
        validFrom:
          type: string
          format: date-time
          example: '2026-01-01T00:00:00+01:00'
        validTo:
          type: string
          format: date-time
          nullable: true
        monday:
          type: integer
          example: 480
        tuesday:
          type: integer
          example: 480
        wednesday:
          type: integer
          example: 480
        thursday:
          type: integer
          example: 480
        friday:
          type: integer
          example: 480
        saturday:
          type: integer
          example: 0
        sunday:
          type: integer
          example: 0
...
    ScheduleTemplateDto:
      type: object
      properties:
        id:
          type: string
          example: 'st-fulltime-40'
        name:
          type: string
          example: '100% Full-Time 40h'
        monday:
          type: integer
          example: 480
        tuesday:
          type: integer
          example: 480
        wednesday:
          type: integer
          example: 480
        thursday:
          type: integer
          example: 480
        friday:
          type: integer
          example: 480
        saturday:
          type: integer
          example: 0
        sunday:
          type: integer
          example: 0
    ScheduleTemplateCreateDto:
      type: object
      required:
        - name
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
      properties:
        name:
          type: string
          example: '100% Full-Time 40h'
        monday:
          type: integer
          example: 480
        tuesday:
          type: integer
          example: 480
        wednesday:
          type: integer
          example: 480
        thursday:
          type: integer
          example: 480
        friday:
          type: integer
          example: 480
        saturday:
          type: integer
          default: 0
          example: 0
        sunday:
          type: integer
          default: 0
          example: 0
    PagedScheduleTemplateDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleTemplateDto'
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 4
        size:
          type: integer
          example: 4
...